+++ src/main/java/org/opentripplanner/api/mapping/ItineraryMapper.java
@@ -32,7 +32,6 @@ public class ItineraryMapper {
         api.waitingTime = domain.waitingTimeSeconds;
         api.walkDistance = domain.nonTransitDistanceMeters;
         api.generalizedCost = domain.generalizedCost;
-        api.walkLimitExceeded = domain.nonTransitLimitExceeded;
         api.elevationLost = domain.elevationLost;
         api.elevationGained = domain.elevationGained;
         api.transfers = domain.nTransfers;