+++ src/main/java/org/opentripplanner/api/parameter/QualifiedModeSet.java
@@ -39,6 +39,7 @@ public class QualifiedModeSet implements Serializable {
         StreetMode accessMode = null;
         StreetMode egressMode = null;
         StreetMode directMode = null;
+        StreetMode transferMode = null;
         Set<TransitMode> transitModes = new HashSet<>();
 
         // Set transit modes
@@ -110,20 +111,24 @@ public class QualifiedModeSet implements Serializable {
             switch (requestMode.mode) {
                 case WALK:
                     accessMode = StreetMode.WALK;
+                    transferMode = StreetMode.WALK;
                     egressMode = StreetMode.WALK;
                     directMode = StreetMode.WALK;
                     break;
                 case BICYCLE:
                     if (requestMode.qualifiers.contains(Qualifier.RENT)) {
                         accessMode = StreetMode.BIKE_RENTAL;
+                        transferMode = StreetMode.BIKE_RENTAL;
                         egressMode = StreetMode.BIKE_RENTAL;
                         directMode = StreetMode.BIKE_RENTAL;
                     } else if (requestMode.qualifiers.contains(Qualifier.PARK)) {
                         accessMode = StreetMode.BIKE_TO_PARK;
+                        transferMode = StreetMode.WALK;
                         egressMode = StreetMode.WALK;
                         directMode = StreetMode.BIKE_TO_PARK;
                     } else {
                         accessMode = StreetMode.BIKE;
+                        transferMode = StreetMode.BIKE;
                         egressMode = StreetMode.BIKE;
                         directMode = StreetMode.BIKE;
                     }
@@ -131,22 +136,27 @@ public class QualifiedModeSet implements Serializable {
                 case CAR:
                     if (requestMode.qualifiers.contains(Qualifier.RENT)) {
                         accessMode = StreetMode.CAR_RENTAL;
+                        transferMode = StreetMode.CAR_RENTAL;
                         egressMode = StreetMode.CAR_RENTAL;
                         directMode = StreetMode.CAR_RENTAL;
                     } else if (requestMode.qualifiers.contains(Qualifier.PARK)) {
                         accessMode = StreetMode.CAR_TO_PARK;
+                        transferMode = StreetMode.WALK;
                         egressMode = StreetMode.WALK;
                         directMode = StreetMode.CAR_TO_PARK;
                     } else if (requestMode.qualifiers.contains(Qualifier.PICKUP)) {
                         accessMode = StreetMode.WALK;
+                        transferMode = StreetMode.WALK;
                         egressMode = StreetMode.CAR_PICKUP;
                         directMode = StreetMode.CAR_PICKUP;
                     } else if (requestMode.qualifiers.contains(Qualifier.DROPOFF)) {
                         accessMode = StreetMode.CAR_PICKUP;
+                        transferMode = StreetMode.WALK;
                         egressMode = StreetMode.WALK;
                         directMode = StreetMode.CAR_PICKUP;
                     } else {
                         accessMode = StreetMode.WALK;
+                        transferMode = StreetMode.WALK;
                         egressMode = StreetMode.WALK;
                         directMode = StreetMode.CAR;
                     }
@@ -169,6 +179,7 @@ public class QualifiedModeSet implements Serializable {
 
         return new RequestModes(
             accessMode,
+            transferMode,
             egressMode,
             directMode,
             transitModes