+++ src/main/java/org/opentripplanner/graph_builder/module/osm/WalkableAreaBuilder.java
@@ -295,8 +295,14 @@ public class WalkableAreaBuilder {
         }
 
         @Override
-        public boolean shouldSkipEdge(Vertex origin, Vertex target, State current, Edge edge,
-                ShortestPathTree spt, RoutingRequest traverseOptions) {
+        public boolean shouldSkipEdge(
+            Set<Vertex> origins,
+            Set<Vertex> targets,
+            State current,
+            Edge edge,
+            ShortestPathTree spt,
+            RoutingRequest traverseOptions
+        ) {
             return !edges.contains(edge);
         }
     }