+++ src/main/java/org/opentripplanner/routing/algorithm/raptor/transit/AccessEgress.java
@@ -2,6 +2,7 @@ package org.opentripplanner.routing.algorithm.raptor.transit;
 
 import org.opentripplanner.model.base.ToStringBuilder;
 import org.opentripplanner.routing.core.State;
+import org.opentripplanner.transit.raptor.api.transit.RaptorCostConverter;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 
 public class AccessEgress implements RaptorTransfer {
@@ -13,20 +14,28 @@ public class AccessEgress implements RaptorTransfer {
 
   private final int durationInSeconds;
 
+  private final int generalizedCost;
+
   /**
    * This should be the last state both in the case of access and egress.
    */
   private final State lastState;
 
-  public AccessEgress(int toFromStop, int durationInSeconds, State lastState) {
+  public AccessEgress(int toFromStop, State lastState) {
     this.toFromStop = toFromStop;
-    this.durationInSeconds = durationInSeconds;
+    this.durationInSeconds = (int) lastState.getElapsedTimeSeconds();
+    this.generalizedCost = RaptorCostConverter.toRaptorCost(lastState.getWeight());
     this.lastState = lastState;
   }
 
   @Override
   public int stop() {
     return toFromStop;
+  }
+
+  @Override
+  public int generalizedCost() {
+    return generalizedCost;
   }
 
   @Override