+++ src/test/java/org/opentripplanner/transit/raptor/_data/stoparrival/Access.java
@@ -1,16 +1,16 @@
 package org.opentripplanner.transit.raptor._data.stoparrival;
 
+import static org.opentripplanner.transit.raptor._data.transit.TestTransfer.walk;
+
 import org.opentripplanner.transit.raptor.api.transit.RaptorTransfer;
 import org.opentripplanner.transit.raptor.api.view.AccessPathView;
-
-import static org.opentripplanner.transit.raptor._data.transit.TestTransfer.walk;
 
 public class Access extends AbstractStopArrival {
 
     private final RaptorTransfer access;
 
     public Access(int stop, int departureTime, int arrivalTime, int cost) {
-        this(stop, arrivalTime, cost, walk(stop, Math.abs(arrivalTime - departureTime)));
+        this(stop, arrivalTime, cost, walk(stop, Math.abs(arrivalTime - departureTime), cost));
     }
 
     public Access(int stop, int arrivalTime, int cost, RaptorTransfer path) {